Creating a list using the first comment #496
WidgetID obtain at first comment.

- delete request from API: /obtain_widget_id
- move fn obtainWidgetId to separate file
- for first comment -> do request to Storage-WRIO-App API for update comment id
